% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\name{alphabet}
\alias{alphabet}
\title{Standard base64 alphabets}
\usage{
alphabet(which = "standard")
}
\arguments{
\item{which}{default \code{"standard"}. Which base64 alphabet to use.}
}
\description{
Standard base64 alphabets
}
\details{
\itemize{
\item \code{"bcrypt"}: the bcrypt alphabet
}

See \href{https://docs.rs/base64/latest/base64/alphabet/index.html#constants}{base64 crate}
for more details.
}
