% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b64-package.R
\docType{package}
\name{b64-package}
\alias{b64}
\alias{b64-package}
\title{b64: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
