% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\alias{decode_file}
\title{Decode base64 encodings}
\usage{
decode(what, eng = engine())

decode_file(path, eng = engine())
}
\arguments{
\item{what}{a scalar character or a raw vector.}
}
\value{
\code{decode()} always returns a \code{blob} object. \code{decode_file()} returns a \code{raw} vector.
}
\description{
Decode base64 encodings
}
