% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{encode}
\alias{encode}
\alias{encode_string}
\alias{encode_raw}
\alias{encode_file}
\title{Encode to base64}
\usage{
encode_string(input)

encode_raw(input)

encode_file(path)
}
\arguments{
\item{input}{A string, raw vector, or file path.}
}
\description{
Encode to base64
}
