% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{decode}
\alias{decode}
\alias{decode_string}
\alias{decode_raw}
\title{Decode from base64}
\usage{
decode_string(input)

decode_raw(input)
}
\arguments{
\item{input}{A string or raw vector.}
}
\description{
Decode from base64
}
