% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\alias{encode_file}
\title{Encode as base64}
\usage{
encode(what, eng = engine())

encode_file(path, eng = engine())
}
\arguments{
\item{what}{a scalar character or a raw vector.}

\item{engine}{a base64 engine. See \code{\link[=engine]{engine()}} for details.}
}
\value{
\code{encode()} is vectorized and will return a character vector of the same
lenght as \code{what}.
}
\description{
Perform base 64 encoding.
}
